Fix string comparisons
When testing whether the first word of a line in the BDF file matches a
specific string, the word is treated as matching as long as it is a
prefix of the string. For example, when testing whether a line begins
with "CAP_HEIGHT", if the first word of the line is simply "CA", this is
considered a match, because only the first two characters are checked.

This is a problem because some BDF files do contain lines that consist
of "CA", as part of their hexadecimal bitmap data. X11's standard 8x13
font is one such case. These lines will be incorrectly treated as
CAP_HEIGHT lines, and the resulting output will not be well-formed.

This commit modifies the string comparisons so that a full match is
required.